systemd: create a systemdtest package
